abstracted scrape runner and db insertion functions out of server.js
